Move extensions and abilities to their pages
